+++ packages/synthchron-tauri/src/components/react-flow/flowStore/yDoc.ts
@@ -1,0 +1,10 @@
+import { Node, Edge } from 'reactflow'
+import { Doc } from 'yjs'
+
+export const yDoc = new Doc()
+export const yDocState = {
+  nodesMap: yDoc.getMap<Node>('nodes'),
+  edgesMap: yDoc.getMap<Edge>('edges'),
+  metaMap: yDoc.getMap('meta'),
+  processModelType: yDoc.getText('processModelType'),
+}
